Automation Stash
Decided to not pursue this automation longer but im putting the code in a branch. The reasons this automation is no longer part of the project is because it would automate less than 5 minutes of coding and to do anything relevant it would require some more days.

Not really worth automating this behaviour upon analysis

There is some cool code in the implementation so im putting it to a branch to not lose it, mostly the usage of a typescript dynamic key interface.